refactor(table): move MemWAL index metadata helpers into lance-table
load_mem_wal_index_details, open_mem_wal_index, new_mem_wal_index_meta and
update_mem_wal_index_compacted_sstables read and write the MemWAL index's
IndexMetadata entry. Every type they touch already lives in lance-table's
system_index module, so they join the data structures they serialize.

lance::index::mem_wal keeps the dataset-level operations and re-exports the
four helpers at their previous visibility.
